Refactor proto compilation path structure
